+++ lib/features/saved_search/delete/actions.dart
@@ -1,0 +1,17 @@
+class SavedSearchDeleteRequestAction {
+  final String savedSearchId;
+
+  SavedSearchDeleteRequestAction(this.savedSearchId);
+}
+
+class SavedSearchDeleteLoadingAction {}
+
+class SavedSearchDeleteSuccessAction {
+  final String savedSearchId;
+
+  SavedSearchDeleteSuccessAction(this.savedSearchId);
+}
+
+class SavedSearchDeleteFailureAction {}
+
+class SavedSearchDeleteResetAction {}
