+++ lib/features/saved_search/delete/middleware.dart
@@ -1,4 +1,4 @@
-import 'package:pass_emploi_app/features/saved_search/delete/slice/actions.dart';
+import 'package:pass_emploi_app/features/saved_search/delete/actions.dart';
 import 'package:pass_emploi_app/redux/states/app_state.dart';
 import 'package:pass_emploi_app/repositories/saved_search/saved_search_delete_repository.dart';
 import 'package:redux/redux.dart';
