--- lib/features/saved_search/delete/slice/actions.dart
@@ -1,19 +1,0 @@
-abstract class SavedSearchDeleteAction {}
-
-class SavedSearchDeleteRequestAction extends SavedSearchDeleteAction {
-  final String savedSearchId;
-
-  SavedSearchDeleteRequestAction(this.savedSearchId);
-}
-
-class SavedSearchDeleteLoadingAction extends SavedSearchDeleteAction {}
-
-class SavedSearchDeleteFailureAction extends SavedSearchDeleteAction {}
-
-class SavedSearchDeleteSuccessAction extends SavedSearchDeleteAction {
-  final String savedSearchId;
-
-  SavedSearchDeleteSuccessAction(this.savedSearchId);
-}
-
-class SavedSearchDeleteResetAction extends SavedSearchDeleteAction {}
