+++ lib/features/saved_search/delete/state.dart
