+++ lib/redux/states/app_state.dart
@@ -1,6 +1,6 @@
 import 'package:equatable/equatable.dart';
 import 'package:pass_emploi_app/configuration/configuration.dart';
-import 'package:pass_emploi_app/features/saved_search/delete/slice/state.dart';
+import 'package:pass_emploi_app/features/saved_search/delete/state.dart';
 import 'package:pass_emploi_app/models/immersion.dart';
 import 'package:pass_emploi_app/models/immersion_details.dart';
 import 'package:pass_emploi_app/models/offre_emploi.dart';
