+++ lib/redux/store/store_factory.dart
@@ -2,7 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:pass_emploi_app/auth/authenticator.dart';
 import 'package:pass_emploi_app/auth/firebase_auth_wrapper.dart';
 import 'package:pass_emploi_app/crashlytics/crashlytics.dart';
-import 'package:pass_emploi_app/features/saved_search/delete/slice/middleware.dart';
+import 'package:pass_emploi_app/features/saved_search/delete/middleware.dart';
 import 'package:pass_emploi_app/models/immersion.dart';
 import 'package:pass_emploi_app/models/offre_emploi.dart';
 import 'package:pass_emploi_app/models/rendezvous.dart';
