+++ lib/widgets/dialogs/saved_search_delete_dialog.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:matomo/matomo.dart';
-import 'package:pass_emploi_app/features/saved_search/delete/presentation/saved_search_delete_view_model.dart';
-import 'package:pass_emploi_app/features/saved_search/delete/slice/actions.dart';
+import 'package:pass_emploi_app/features/saved_search/delete/actions.dart';
+import 'package:pass_emploi_app/presentation/saved_search/saved_search_delete_view_model.dart';
 import 'package:pass_emploi_app/ui/font_sizes.dart';
 import 'package:pass_emploi_app/widgets/buttons/primary_action_button.dart';
 import 'package:pass_emploi_app/widgets/buttons/secondary_button.dart';
