+++ test/presentation/saved_search/saved_search_delete_view_model_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:pass_emploi_app/features/saved_search/delete/presentation/saved_search_delete_view_model.dart';
-import 'package:pass_emploi_app/features/saved_search/delete/slice/actions.dart';
-import 'package:pass_emploi_app/features/saved_search/delete/slice/state.dart';
+import 'package:pass_emploi_app/features/saved_search/delete/actions.dart';
+import 'package:pass_emploi_app/features/saved_search/delete/state.dart';
+import 'package:pass_emploi_app/presentation/saved_search/saved_search_delete_view_model.dart';
 import 'package:pass_emploi_app/redux/states/app_state.dart';
 
 import '../../doubles/spies.dart';
